[fix]: [PL-30123]: Reverts CompatibleFieldserializer changes for cache (#40840)
* Revert "[fix]:[PL-29719]: Replacing kryo FieldSerailizer with CompatibleFieldSerializer in cache module (#40037)"

This reverts commit 8aedd7431ffa4892e09e4f854530cb20eacf8921.

* [fix]: [PL-30123]: Updates build number to 77614

* [fex]: [PL-30123]: Updates build number to fix PR check
