+++ 840-template-service/src/main/resources/redisson-jcache.yaml
@@ -1,5 +1,5 @@
 singleServerConfig:
   address: "redis://127.0.0.1:6379"
-codec: !<io.harness.redis.CompatibleFieldSerializerWithoutReferenceCodec> {}
+codec: !<io.harness.redis.RedissonKryoCodec> {}
 nettyThreads: 16
 useScriptCache: true
