+++ 960-persistence/src/main/java/io/harness/redis/RedissonKryoCodec.java
@@ -9,7 +9,66 @@ package io.harness.redis;
 
 import static io.harness.annotations.dev.HarnessTeam.PL;
 
+import static java.lang.String.format;
+
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.exception.GeneralException;
+import io.harness.reflection.CodeUtils;
+import io.harness.reflection.HarnessReflections;
+import io.harness.serializer.ClassResolver;
+import io.harness.serializer.HKryo;
+import io.harness.serializer.KryoRegistrar;
+import io.harness.serializer.KryoSerializer;
+
+import com.esotericsoftware.kryo.Kryo;
+import com.esotericsoftware.kryo.Registration;
+import com.esotericsoftware.kryo.Serializer;
+import com.esotericsoftware.kryo.serializers.FieldSerializer;
+import com.esotericsoftware.kryo.util.IntMap;
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
+import java.util.List;
+import java.util.Set;
+import org.redisson.codec.KryoCodec;
 
 @OwnedBy(PL)
-public class RedissonKryoCodec extends AbstractRedissonKryoCodec {}
+public class RedissonKryoCodec extends KryoCodec {
+  @Override
+  protected Kryo createInstance(List<Class<?>> classes, ClassLoader classLoader) {
+    return kryo();
+  }
+
+  private synchronized Kryo kryo() {
+    final ClassResolver classResolver = new ClassResolver();
+    HKryo kryo = new HKryo(classResolver);
+    kryo.setDefaultSerializer(getDefaultSerializer());
+
+    try {
+      Set<Class<? extends KryoRegistrar>> kryoRegistrars = HarnessReflections.get().getSubTypesOf(KryoRegistrar.class);
+      for (Class<? extends KryoRegistrar> clazz : kryoRegistrars) {
+        Constructor<?> constructor = clazz.getConstructor();
+        final KryoRegistrar kryoRegistrar = (KryoRegistrar) constructor.newInstance();
+
+        final IntMap<Registration> previousState = new IntMap<>(classResolver.getRegistrations());
+        kryo.setCurrentLocation(CodeUtils.location(kryoRegistrar.getClass()));
+        kryoRegistrar.register(kryo);
+
+        try {
+          KryoSerializer.check(previousState, classResolver.getRegistrations());
+        } catch (Exception exception) {
+          throw new IllegalStateException(
+              format("Check for registration of %s failed", clazz.getCanonicalName()), exception);
+        }
+      }
+
+    } catch (NoSuchMethodException | IllegalAccessException | InstantiationException | InvocationTargetException e) {
+      throw new GeneralException("Failed initializing kryo", e);
+    }
+
+    return kryo;
+  }
+
+  protected Class<? extends Serializer> getDefaultSerializer() {
+    return FieldSerializer.class;
+  }
+}
